This is alot
So i tried to fix the associations in the models, so the foreign keys would actually connect to the right tabels. (Still dosent work, but im getting somewhere)

Then i changed the faker data output to see and test where the data would change.

Then i got to my frontend and fetched data from localhost in the server, and the server is getting the data from the database.